basic data covered by tests
