+++ BodyweightFitness/Domain/WorkoutChartType.swift
@@ -1,0 +1,4 @@
+enum WorkoutChartType {
+    case WorkoutLength
+    case CompletionRate
+}
