+++ BodyweightFitnessTests/Domain/DataEntriesCompanionSpec.swift
@@ -24,7 +24,8 @@ class DataEntriesCompanionSpec: QuickSpec {
                     let dataEntries = dataEntriesCompanion.getDataEntries(
                             fromDate: fromDate,
                             numberOfDays: 7,
-                            repositoryRoutines: []
+                            repositoryRoutines: [],
+                            workoutChartType: .WorkoutLength
                     )
 
                     expect(dataEntries.count).to(equal(7))
@@ -46,7 +47,8 @@ class DataEntriesCompanionSpec: QuickSpec {
                     let dataEntries = dataEntriesCompanion.getDataEntries(
                             fromDate: fromDate,
                             numberOfDays: 30,
-                            repositoryRoutines: []
+                            repositoryRoutines: [],
+                            workoutChartType: .WorkoutLength
                     )
 
                     expect(dataEntries.count).to(equal(30))
@@ -59,6 +61,35 @@ class DataEntriesCompanionSpec: QuickSpec {
 
                     expect(dataEntries[29].x).to(equal(29))
                     expect(dataEntries[29].y).to(equal(0))
+                }
+
+                it("returns data entries for last 7 days with logged workouts") {
+                    let fromDate = self.mockDate(from: "2017-12-10T13:00:00Z")
+
+                    let firstRepositoryRoutine = RepositoryRoutine()
+                    firstRepositoryRoutine.startTime = self.mockDate(from: "2017-12-09T12:00:00")
+                    firstRepositoryRoutine.lastUpdatedTime = self.mockDate(from: "2017-12-09T12:50:00")
+
+                    let dataEntriesCompanion = DataEntriesCompanion()
+                    let dataEntries = dataEntriesCompanion.getDataEntries(
+                            fromDate: fromDate,
+                            numberOfDays: 7,
+                            repositoryRoutines: [
+                                firstRepositoryRoutine
+                            ],
+                            workoutChartType: .WorkoutLength
+                    )
+
+                    expect(dataEntries.count).to(equal(7))
+
+                    expect(dataEntries[0].x).to(equal(0))
+                    expect(dataEntries[0].y).to(equal(0))
+
+                    expect(dataEntries[1].x).to(equal(1))
+                    expect(dataEntries[1].y).to(equal(50))
+
+                    expect(dataEntries[6].x).to(equal(6))
+                    expect(dataEntries[6].y).to(equal(0))
                 }
             }
         }
